pyproject.toml
[tool.poetry]
name = "mongo_python"
version = "25.0.dev"
description = ""
authors = ["ADTRAN"]

[tool.poetry.dependencies]
python = "^3.8"
xmltodict = "^0.12.0"
pexpect = "^4.8.0"
requests = "^2.10.0"

[[tool.poetry.source]]
name = "artifactory"
url = "https://artifactory.adtran.com/artifactory/api/pypi/pypi/simple"
default = true

[tool.poetry.scripts]
mongo-post = "mongo_python.mongo_create:run_cli"
mongo = "mongo_python.mongo_update:run_cli"

[tool.poetry.dev-dependencies]
pytest = "^5.0"
pylama = "8.4.1"
pylint = "^2.2"
pytest-cov = "^2.6"
pyhamcrest = "^2.0.0"
httpretty = "1.1.4"
mock = "^3.0.0"
black = "18.9b0"
sphinx = "^2.0"
pyflakes = "3.0.1"
requests = "^2.10.0"

[tool.porg]
default_python_version = "3.8"
base_docker_image = "artifactory.adtran.com:443/docker_adtran-local/porg-base-image-38:3.3.6"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
fail_under = 80
exclude_lines = [
    "if __name__ == .__main__.:",
    "raise NotImplementedError()",
]
